requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50
langgraph>=0.0.30

# Vector store and embeddings
faiss-cpu>=1.7.4
openai>=1.0.0
tiktoken>=0.5.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0

# Additional dependencies
ipython>=8.0.0
matplotlib>=3.7.0
pillow>=9.0.0
python-dotenv>=1.0.0

# Optional but recommended for better performance
scikit-learn>=1.3.0
scipy>=1.10.0

# influxdb-client for InfluxDB integration
influxdb-client>=1.35.0